requirements.txt
awesome-slugify==1.6.5
beautifulsoup4==4.8.1
bleach==2.1.3
bootstrap4==0.1.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
Django==2.0
django-appconf==1.0.3
django-bootstrap-modal-forms==1.4.4
django-bootstrap4==1.0.1
django-braces==1.13.0
django-channels==0.7.0
django-compressor==2.2
django-datatables-view==1.19.0
django-imagekit==4.0.2
django-statici18n==1.7.0
django-template-forms==0.3.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
html5lib==1.0.1
idna==2.8
lib==3.0.0
libsass==0.14.0
lxml==4.4.1
Markdown==2.6.11
markdown2==2.3.5
mysqlclient==1.4.4
oauthlib==3.1.0
pilkit==2.0
Pillow==6.2.1
premailer==3.1.1
psycopg2==2.8.4
pytz==2019.3
rcssmin==1.0.6
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
six==1.12.0
soupsieve==1.9.4
Unidecode==0.4.21
urllib3==1.25.6
webencodings==0.5.1
